move class to impl
