+++ src/main/java/youyihj/zenutils/impl/fluid/MCLiquidTankProperties.java
@@ -1,10 +1,11 @@
-package youyihj.zenutils.api.liquid;
+package youyihj.zenutils.impl.fluid;
 
 import crafttweaker.annotations.ZenRegister;
 import crafttweaker.api.liquid.ILiquidStack;
 import crafttweaker.api.minecraft.CraftTweakerMC;
 import net.minecraftforge.fluids.capability.IFluidTankProperties;
 import stanhebben.zenscript.annotations.ZenClass;
+import youyihj.zenutils.api.liquid.ILiquidTankProperties;
 
 /**
  * @author various.authors
